refactor: feature flag typing remove key-value
Key-value allows just about anything and it is not that great for
typing. This change removes the key-value part of the FeatureFlag and
forces all apps to strongly type it. We can reintroduce it when we
absolutely need it.
