+++ tensorboard/webapp/feature_flag/actions/BUILD
@@ -8,7 +8,6 @@ tf_ng_module(
         "feature_flag_actions.ts",
     ],
     deps = [
-        "//tensorboard/webapp/feature_flag:types",
         "//tensorboard/webapp/feature_flag/store:types",
         "@npm//@ngrx/store",
     ],
