+++ tensorboard/webapp/feature_flag/testing.ts
@@ -1,4 +1,4 @@
-/* Copyright 2020 The TensorFlow Authors. All Rights Reserved.
+/* Copyright 2019 The TensorFlow Authors. All Rights Reserved.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -12,18 +12,15 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
-import {Injectable} from '@angular/core';
 
-import {FeatureFlags} from '../feature_flag/types';
+import {FeatureFlags} from './types';
 
-@Injectable()
-export abstract class TBFeatureFlagDataSource {
-  /**
-   * Gets feature flags defined.
-   *
-   * The "feature" is very loosely defined so other applications can define more
-   * flags. It is up to the application to better type the flags and create necessary
-   * facilities (e.g., strongly typed selector).
-   */
-  abstract getFeatures(): FeatureFlags;
+export function buildFeatureFlag(
+  override: Partial<FeatureFlags> = {}
+): FeatureFlags {
+  return {
+    enabledExperimentalPlugins: [],
+    inColab: false,
+    ...override,
+  };
 }
