+++ tensorboard/webapp/webapp_data_source/tb_feature_flag_testing.ts
@@ -22,6 +22,7 @@ export class TestingTBFeatureFlagDataSource extends TBFeatureFlagDataSource {
   getFeatures() {
     return {
       enabledExperimentalPlugins: [] as string[],
+      inColab: false,
     };
   }
 }
